requirements/requirements.txt
############################################
# Dependencies
############################################
schedule>=1.1.0,<1.2
webdriver-manager>=3.8.3,<3.9
undetected-chromedriver>=3.1.5,<3.2
python-dotenv>=0.21.0,<0.30
